Refactor: tpl단위로 파일별로 나눔
