+++ src/js/App.js
@@ -1,8 +1,13 @@
 import React, { Component } from 'react';
+
 import '../scss/app.css';
-import { MakeAddTodoTpl, MakeTodoListContainerTpl } from './make-tpl.js';
-import { getData, taskDataUrl } from './fetch-data.js';
+
+import AddTodoTpl from './template/add-todo-tpl';
+import TodoListContainerTpl from './template/todo-list-container-tpl.js'
 import StatusBoard from './status-board.js';
+
+import { getData, taskDataUrl } from './fetch-data.js';
+
 
 class App extends Component {
   constructor(props) {
@@ -58,12 +63,12 @@ class App extends Component {
     return (
       <div className="todo-app-conatiner">
         <StatusBoard tasks={[...this.state.tasks]} />
-        <MakeAddTodoTpl
+        <AddTodoTpl
           taskState={this.state}
           initTask={this.initTask}
           handleChangeWord={this.handleChangeWord}
         />
-        <MakeTodoListContainerTpl
+        <TodoListContainerTpl
           taskState={this.state}
           handleBFolded={this.handleBFolded}
           handleBTodo={this.handleBTodo}
