+++ src/js/inputer.js
@@ -10,7 +10,7 @@ class Inputer extends Component {
   };
 
   addTask = e => {
-    const tasks = [...this.props.taskState.tasks]
+    const tasks = [...this.props.taskState.tasks];
     let IDNumber = this.props.taskState.IDNum;
     if (tasks.some(v => v.title === this.props.taskState.word)) return;
     tasks.push({
@@ -25,12 +25,12 @@ class Inputer extends Component {
     return (
       <span>
         <input
-          className='add-todo-inputer'
+          className="add-todo-inputer"
           value={this.props.taskState.word}
           onChange={this.handleChange}
           onKeyPress={this.handleEnterKeyPress}
         />
-        <button className='add-todo-inputer-button' onClick={this.addTask}>
+        <button className="add-todo-inputer-button" onClick={this.addTask}>
           입력
         </button>
       </span>
