+++ src/js/template/fold-btn-tpl.js
@@ -1,10 +1,10 @@
 import React from 'react';
 
-const FoldTask = props => {
+const FoldBtnTpl = props => {
   return (
     <button className={props.className} onClick={props.handleBFolded}>
       접기
     </button>
   );
 };
-export default FoldTask;
+export default FoldBtnTpl;
