+++ src/js/template/loading-tpl.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const LoadingTpl = props => {
+  return <div className="loading" />;
+};
+
+export default LoadingTpl
