+++ src/js/template/nav-btn-tpl.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const NavBtnTpl = props => {
+  return (
+    <button className="todo-list-nav-button" onClick={props.handleBTodo}>
+      {props.taskState.bTodo ? '할 일' : '한 일'}
+    </button>
+  );
+};
+
+export default NavBtnTpl
